+++ include/build_config/SDL_build_config_minimal.h
@@ -98,4 +98,7 @@ typedef unsigned int uintptr_t;
 /* Enable dialog subsystem */
 #define SDL_DIALOG_DUMMY 1
 
+/* Enable tray subsystem */
+#define SDL_TRAY_DUMMY 1
+
 #endif /* SDL_build_config_minimal_h_ */
