+++ include/build_config/SDL_build_config_xbox.h
@@ -220,4 +220,7 @@
 /* Enable dialog subsystem */
 #define SDL_DIALOG_DUMMY 1
 
+/* Enable tray subsystem */
+#define SDL_TRAY_DUMMY 1
+
 #endif /* SDL_build_config_wingdk_h_ */
