+++ src/tray/dummy/SDL_tray.c
@@ -21,7 +21,7 @@
 
 #include "SDL_internal.h"
 
-#ifndef SDL_PLATFORM_MACOS
+#ifdef SDL_TRAY_DUMMY
 
 #include "../SDL_tray_utils.h"
 
@@ -140,4 +140,4 @@ void SDL_DestroyTray(SDL_Tray *tray)
 {
 }
 
-#endif // !SDL_PLATFORM_MACOS
+#endif // SDL_TRAY_DUMMY
